Refactor docker_container and docker_container_network rspec tests
