+++ spec/libraries/container_networks_spec.rb
@@ -1,20 +1,18 @@
 require 'spec_helper'
 require 'docker'
-require_relative '../../libraries/docker_base'
-require_relative '../../libraries/docker_container'
+require_relative '../../libraries/helpers_network'
 
-describe DockerCookbook::DockerContainer do
-  let(:resource) { DockerCookbook::DockerContainer.new('hello_world') }
-
-  it 'has a default action of [:run]' do
-    expect(resource.action).to eql([:run])
-  end
+describe 'docker_container' do
+  step_into :docker_container
+  platform 'ubuntu'
 
   describe 'gets ip_address_from_container_networks' do
+    include DockerCookbook::DockerHelpers::Network
     let(:options) { { 'id' => rand(10_000).to_s } }
     subject do
       Docker::Container.send(:new, Docker.connection, options)
     end
+
     # https://docs.docker.com/engine/api/version-history/#v121-api-changes
     context 'when docker API < 1.21' do
       let(:ip_address) { '10.0.0.1' }
@@ -24,12 +22,14 @@ describe DockerCookbook::DockerContainer do
           'IPAddress' => ip_address,
         }
       end
+
       it 'gets ip_address as nil' do
-        actual = resource.ip_address_from_container_networks(subject)
-        expect { resource.ip_address_from_container_networks(subject) }.not_to raise_error
+        actual = ip_address_from_container_networks(subject)
+        expect { ip_address_from_container_networks(subject) }.not_to raise_error
         expect(actual).to eq(nil)
       end
     end
+
     context 'when docker API > 1.21' do
       let(:ip_address) { '10.0.0.1' }
       let(:options) do
@@ -46,8 +46,9 @@ describe DockerCookbook::DockerContainer do
           },
         }
       end
+
       it 'gets ip_address' do
-        actual = resource.ip_address_from_container_networks(subject)
+        actual = ip_address_from_container_networks(subject)
         expect(actual).to eq(ip_address)
       end
     end
